Project/requirements.txt
absl-py==1.0.0
apache-beam==2.37.0
asgiref==3.4.1
astunparse==1.6.3
avro-python3==1.10.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
dill==0.3.1.1
Django==3.2.8
djangorestframework==3.12.4
dm-tree==0.1.6
docopt==0.6.2
fastavro==1.4.10
flatbuffers==2.0
fonttools==4.31.1
gast==0.5.3
gin-config==0.5.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
h5py==3.6.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
joblib==1.1.0
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
mssql-django==1.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
object-detection @ file:///C:/Users/Jc/Desktop/FYP/models/research
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.7
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
portalocker==2.4.0
promise==2.3
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pymongo==3.12.3
pyodbc==4.0.32
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
pywin32==303
PyYAML==5.4.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.1
termcolor==1.1.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tqdm==4.63.0
typeguard==2.13.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
